Move Mapping code to its own mod.
